+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,71 +1,12 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
 
-public class UserService {
+public interface UserService {
 
-    private final DataSource dataSource;
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findById(final long id);
 
-    public UserService(final DataSource dataSource, final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.dataSource = dataSource;
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        executeSameConnection(() -> {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-        });
-    }
-
-    private void executeSameConnection(Runnable runnable) {
-        Connection connection = null;
-        try {
-            connection = DataSourceUtils.getConnection(dataSource);
-            connection.setAutoCommit(false);
-            runnable.run();
-            connection.commit();
-        } catch (RuntimeException e) {
-            rollback(connection, e);
-            throw e;
-        } catch (SQLException e) {
-            rollback(connection, e);
-            throw new DataAccessException(e);
-        } finally {
-            try {
-                connection.setAutoCommit(true);
-            } catch (SQLException ignored) {
-            }
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
-
-    private void rollback(Connection connection, Exception e) {
-        try {
-            connection.rollback();
-        } catch (SQLException ex) {
-            e.addSuppressed(ex);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
